geoapi/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.2.10; python_version >= '3.5'
django-cors-headers==3.10.1
django==3.1
djangorestframework==3.12
pytz==2021.3
sqlparse==0.4.2; python_version >= '3.5'
